Content  pseudo-element  isn't displaying  correctly on page.
In cases where the content: '-' is not working, it's possible that the ::before pseudo-element is not being properly positioned on top of the parent element. One possible way to solve this issue is to add the position: relative rule to the parent element. For instance, we can add the position: relative rule to the .accordion-label class, which would become: